Contracts logic haschange and commited

+++ src/com/oganalysis/business/ContractsBusinessService.java
@@ -6,13 +6,13 @@ import java.util.Map;
 public interface ContractsBusinessService {
 	
 		
-	public Map<String,Map<Integer,Float>> getQuantityByCompany(Map<String,List<String>> selectedOptions,int startDate,int endDate);
-	public Map<String,Map<Integer,Float>> getQuantityByCountry(Map<String,List<String>> selectedOptions,int startDate,int endDate);
-	public Map<String,Map<Integer,Float>> getQuantityByTerminal(Map<String,List<String>> selectedOptions,int startDate,int endDate);
+	public Map<String,Map<Integer,Double>> getQuantityByCompany(Map<String,List<String>> selectedOptions,int startDate,int endDate);
+	public Map<String,Map<Integer,Double>> getQuantityByCountry(Map<String,List<String>> selectedOptions,int startDate,int endDate);
+	public Map<String,Map<Integer,Double>> getQuantityByTerminal(Map<String,List<String>> selectedOptions,int startDate,int endDate);
 	
 	public List<String> getImportCountries(List<String> exportCountries);
 	public List<String> getImportCompanies(List<String> exportCompanies);
 	public List<String> getExportCompanies();
 	
-	public Map<String,Map<Integer,Float>> getQuantitiesForRecord(Map<String,List<String>>selectedOptions,int startDate, int endDate,String displayType,String recordName);
+	public Map<String,Map<Integer,Double>> getQuantitiesForRecord(Map<String,List<String>>selectedOptions,int startDate, int endDate,String displayType,String recordName);
 }
